Implement proper error handling for out of memory conditions
Make sure that memory is available by calling posix_fallocate() on Unix.
Otherwise memory will only be allocated when it is actually accessed and
an out of memory condition will cause a fatal SIGBUS signal.

Use exceptions for handling out of memory conditions and make sure that
the exceptions are handled somewhere.