Make sure callbacks from workers are executed on the correct thread
When using the QObject::connect() call without a second object argument,
the functor will be executed on the thread were the signal is issued (i.e.
in the worker thread). Pass this as a second object to make sure it gets
executed on the GUI thread (because it will access the object).