Download most build dependencies during Windows CI build
Download build dependencies (except VS2015 and Qt) during CI build in order
to reduce the amount of configuration needed on the build server.
VS2015 is not downloaded automatically because it is non-free, Qt is not
downloaded automatically because it is too large and installation is to
difficult.

Downloaded files are stored into tools/build-dep/, if the files already
exist they will not be downloaded.
For every file to be downloaded there is a .uri file in tools/build-dep/
(which contains the URI the file was downloaded from originally) and a
.sha512sum file (which contains the SHA512 sum which is used to check the
integrity of the file). This directory also contains URIs and checksums for
the sources for e.g. HDF5 which are not needed and not downloaded.
Normally files will not be downloaded from the original location, but
instead from a mirror whose URI can be specified using the
VOXIEBUILD_URL_BASE environment variable or the URL_BASE=... command line
parameter. You can use https://www.ipvs.uni-stuttgart.de/files/pas/src as a
mirror.